refactor: rename ssh_server -> gateway, add GatewayServer/TerminalProxy/CmdProxy, split handlers, fix exec no-TTY issue, add enable_ssh_gateway toggle
- Rename ssh_server/ module to gateway/ for consistency
- Rename SSHAgentServer -> GatewayServer, NATSTerminal -> TerminalProxy,
  SSHExec -> CmdProxy, SSHSessionHandler -> DirectSessionHandler
- Split handlers.py (831 lines) into handlers.py (DirectSessionHandler)
  and menu.py (MenuSessionHandler)
- Move SSHSessionAuditSerializer from inline in models.py to serializers.py
- Remove SSH_EXEC_HANDOVER.md (deprecated handoff doc)
- Fix exec issue: SSH 'ssh host command' with no TTY no longer requires
  stdin to keep channel open; connection_lost defers cleanup until
  NATS response arrives via _exec_completed flag
- Add enable_ssh_gateway field to CoreSettings + migration 0054
- Gate run_ssh_server management command on core.ssh_gateway_enabled
- Expose ssh_gateway_enabled in dashinfo API
- Add permission re-check in MenuSessionHandler._connect_to_agent()
- Create AGENTS.md with dev workflow documentation
