+++ .gitignore
@@ -61,3 +61,4 @@ query_schema.json
 gunicorn_config.py
 set_local_config.py
 api/tacticalrmm/tacticalrmm/private/exe/mesh*
+AGENTS.md
