+++ api/tacticalrmm/tacticalrmm/gateway/egg.py
