+++ api/tacticalrmm/tacticalrmm/gateway/menu.py
@@ -8,11 +8,9 @@ from django.conf import settings
 from django.utils import timezone as djangotime
 
 from .audit import _close_session_and_audit, _record_session_and_audit
-from .exec import SSHExec
-from .highscores import _add_highscore, _format_highscores
-from .terminal import NATSTerminal
-from .utils import _get_user_group, _resolve_and_check
 from .egg import EggGame
+from .terminal import TerminalProxy
+from .utils import _get_user_group, _resolve_and_check
 
 logger = logging.getLogger("trmm")
 
@@ -50,196 +48,6 @@ def _get_menu_agents(user):
             (a.agent_id, a.hostname, a.public_ip or "", a.status, a.version, a.last_seen)
         )
     return tree
-
-
-class SSHSessionHandler(asyncssh.SSHServerSession):
-    def __init__(self, user, agent, session_id, remote_ip,
-                 client_version="", ssh_key_name="", ssh_key_type="",
-                 ssh_key_fingerprint=""):
-        super().__init__()
-        self._user = user
-        self._agent = agent
-        self._session_id = session_id
-        self._remote_ip = remote_ip
-        self._client_version = client_version
-        self._ssh_key_name = ssh_key_name
-        self._ssh_key_type = ssh_key_type
-        self._ssh_key_fingerprint = ssh_key_fingerprint
-        self._term = None
-        self._exec = None
-        self._chan = None
-        self._started_at = None
-        self._terminal_type = ""
-        self._terminal_rows = 0
-        self._terminal_cols = 0
-        self._session_type = None
-        self._exec_cmd = None
-
-    def connection_made(self, chan):
-        self._chan = chan
-        try:
-            peer_name = chan.get_extra_info("peername", ("", ""))
-            self._remote_ip = peer_name[0] if peer_name else self._remote_ip
-            self._started_at = djangotime.now()
-            logger.info(
-                "SSH connection made user=%s agent=%s remote_ip=%s client=%s",
-                self._user.username, self._agent.agent_id, self._remote_ip, self._client_version,
-            )
-        except Exception as e:
-            logger.error("SSH connection_made failed: %s", e, exc_info=True)
-            raise
-
-    def exec_requested(self, command):
-        self._session_type = "exec"
-        self._exec_cmd = command
-        logger.info(
-            "SSH exec requested user=%s agent=%s command=%s",
-            self._user.username, self._agent.agent_id, command,
-        )
-        asyncio.create_task(self._start_exec())
-        return True
-
-    async def _start_exec(self):
-        async def output_cb(data, done=False, exit_code=None):
-            try:
-                if isinstance(data, bytes):
-                    data = data.decode("utf-8", errors="replace")
-                if self._chan and not self._chan.is_closing():
-                    self._chan.write(data)
-                if done:
-                    asyncio.create_task(_close_session_and_audit(
-                        self._user, self._agent, self._session_id, self._remote_ip,
-                        self._started_at,
-                    ))
-                    self._chan.exit(exit_code or 0)
-            except Exception:
-                logger.error("SSH exec output_cb error", exc_info=True)
-
-        try:
-            self._exec = SSHExec(self._agent, self._session_id, self._exec_cmd)
-            await self._exec.start(output_cb)
-        except Exception as e:
-            logger.error("SSH exec start failed: %s", e, exc_info=True)
-            await self._chan.write(f"\r\nFailed to execute command: {e}\r\n")
-            self._chan.exit(1)
-
-    def shell_requested(self):
-        self._session_type = "shell"
-        shell = self._agent.effective_default_shell
-        self._term = NATSTerminal(self._agent, self._session_id, shell)
-        role_name = "None"
-        if self._user.role:
-            role_name = self._user.role.name
-        self._chan.write(
-            f"\r\n\x1b[32mWelcome, \x1b[1m{self._user.username}\x1b[0m\x1b[32m [Role: {role_name}]\x1b[0m\r\n\r\n"
-        )
-        asyncio.create_task(self._start_terminal())
-        asyncio.create_task(
-            _record_session_and_audit(
-                self._user, self._agent, self._session_id, self._remote_ip,
-                client_version=self._client_version,
-                ssh_key_name=self._ssh_key_name,
-                ssh_key_type=self._ssh_key_type,
-                ssh_key_fingerprint=self._ssh_key_fingerprint,
-            )
-        )
-        logger.info(
-            "SSH shell started user=%s agent=%s remote_ip=%s client=%s",
-            self._user.username, self._agent.agent_id, self._remote_ip, self._client_version,
-        )
-        return True
-
-    async def _start_terminal(self):
-        async def output_cb(data, done=False, exit_code=None):
-            try:
-                if isinstance(data, bytes):
-                    data = data.decode("utf-8", errors="replace")
-                if self._chan and not self._chan.is_closing():
-                    self._chan.write(data)
-                if done:
-                    asyncio.create_task(_close_session_and_audit(
-                        self._user, self._agent, self._session_id, self._remote_ip,
-                        self._started_at,
-                        terminal_type=self._terminal_type,
-                        terminal_rows=self._terminal_rows,
-                        terminal_cols=self._terminal_cols,
-                    ))
-                    self._chan.exit(exit_code or 0)
-            except Exception:
-                logger.error("SSH terminal output_cb error", exc_info=True)
-
-        try:
-            await self._term.start(output_cb)
-        except Exception as e:
-            logger.error("SSH terminal start failed: %s", e, exc_info=True)
-            await self._chan.write(f"\r\nFailed to start terminal: {e}\r\n")
-            self._chan.exit(1)
-
-    def pty_requested(self, term_type, term_size, term_modes):
-        self._terminal_type = term_type
-        if term_size:
-            self._terminal_cols, self._terminal_rows = term_size[0], term_size[1]
-        return True
-
-    def terminal_modes(self):
-        return {
-            asyncssh.VEOF: 4,
-            asyncssh.VINTR: 3,
-            asyncssh.VKILL: 21,
-            asyncssh.VQUIT: 28,
-            asyncssh.VSTART: 17,
-            asyncssh.VSTOP: 19,
-            asyncssh.VSUSP: 26,
-            asyncssh.VTIME: 0,
-            asyncssh.VMIN: 1,
-            asyncssh.ECHO: 0,
-            asyncssh.ECHOE: 0,
-            asyncssh.ECHOK: 0,
-            asyncssh.ECHOKE: 0,
-            asyncssh.ECHOCTL: 0,
-            asyncssh.ECHOPRT: 0,
-            asyncssh.ISIG: 1,
-            asyncssh.ICANON: 1,
-            asyncssh.IEXTEN: 1,
-            asyncssh.CTERMINAL: 0,
-        }
-
-    def data_received(self, data, datatype):
-        if self._term:
-            asyncio.create_task(self._term.write(data))
-
-    def connection_lost(self, exc):
-        if self._term:
-            asyncio.create_task(self._term.stop())
-        if self._exec:
-            asyncio.create_task(self._exec.stop())
-        if exc:
-            logger.error("SSH connection lost: %s", exc)
-
-    def terminal_size_changed(self, w, h, pw, ph):
-        if self._term:
-            asyncio.create_task(self._term.resize(h, w))
-
-    def closed(self):
-        if self._term:
-            asyncio.create_task(self._term.stop())
-        if self._exec:
-            asyncio.create_task(self._exec.stop())
-        if self._started_at and self._session_type == "shell":
-            asyncio.create_task(
-                _close_session_and_audit(
-                    self._user, self._agent, self._session_id, self._remote_ip,
-                    self._started_at,
-                    terminal_type=self._terminal_type,
-                    terminal_rows=self._terminal_rows,
-                    terminal_cols=self._terminal_cols,
-                )
-            )
-            logger.info(
-                "SSH session ended user=%s agent=%s duration=%ds",
-                self._user.username, self._agent.agent_id,
-                int((djangotime.now() - self._started_at).total_seconds()),
-            )
 
 
 class MenuSessionHandler(asyncssh.SSHServerSession):
@@ -308,7 +116,6 @@ class MenuSessionHandler(asyncssh.SSHServerSession):
     def data_received(self, data, datatype):
         try:
             if self._state == "terminal":
-                # Check for return-to-menu key (Ctrl+^ = 0x1e)
                 if isinstance(data, bytes):
                     if b'\x1e' in data:
                         asyncio.create_task(self._return_from_terminal())
@@ -341,13 +148,13 @@ class MenuSessionHandler(asyncssh.SSHServerSession):
             for ch in text:
                 asyncio.create_task(self._handle_char(ch))
         except Exception as e:
-            logger.error("SSH menu data_received error: %s", e, exc_info=True)
+            logger.error("Gateway menu data_received error: %s", e, exc_info=True)
 
     def connection_lost(self, exc):
         if self._term:
             asyncio.create_task(self._term.stop())
         if exc:
-            logger.error("SSH menu connection lost: %s", exc)
+            logger.error("Gateway menu connection lost: %s", exc)
 
     def terminal_size_changed(self, w, h, pw, ph):
         if self._term:
@@ -381,7 +188,7 @@ class MenuSessionHandler(asyncssh.SSHServerSession):
         msg = f"\r\n\x1b[32mWelcome, \x1b[1m{self._user.username}\x1b[0m\x1b[32m [Role: {group}]\x1b[0m\r\n"
         await self._write(msg)
         logger.info(
-            "SSH menu session started user=%s remote_ip=%s",
+            "Gateway menu session started user=%s remote_ip=%s",
             self._user.username, self._remote_ip,
         )
         await self._show_clients()
@@ -391,7 +198,7 @@ class MenuSessionHandler(asyncssh.SSHServerSession):
         clients = sorted(self._tree.keys())
         lines = [
             "",
-            f"\x1b[1mTactical RMM SSH Gateway\x1b[0m \x1b[2mv{settings.TRMM_VERSION}\x1b[0m",
+            f"\x1b[1mTactical RMM Gateway\x1b[0m \x1b[2mv{settings.TRMM_VERSION}\x1b[0m",
             "\x1b[2mSelect a client to browse its agents\x1b[0m",
             "",
         ]
@@ -579,10 +386,19 @@ class MenuSessionHandler(asyncssh.SSHServerSession):
                 self._chan.write(ch)
                 return
         except Exception as e:
-            logger.error("SSH menu input error: %s", e, exc_info=True)
+            logger.error("Gateway menu input error: %s", e, exc_info=True)
 
     async def _connect_to_agent(self, agent_id, hostname):
         from agents.models import Agent
+
+        agent = await _resolve_and_check(self._user, agent_id)
+        if agent is None:
+            await self._write(
+                f"\r\n\x1b[31mAccess denied: you don't have permission "
+                f"to access agent {hostname}\x1b[0m\r\n"
+            )
+            return
+
         try:
             agent = await sync_to_async(
                 lambda: Agent.objects.get(agent_id=agent_id)
@@ -594,7 +410,7 @@ class MenuSessionHandler(asyncssh.SSHServerSession):
         self._selected_agent = agent
         self._state = "terminal"
         shell = agent.effective_default_shell
-        self._term = NATSTerminal(agent, self._session_id, shell)
+        self._term = TerminalProxy(agent, self._session_id, shell)
 
         await self._write(
             f"\r\n\x1b[32mConnecting to {hostname} ({agent_id})...\x1b[0m\r\n\r\n"
@@ -610,7 +426,7 @@ class MenuSessionHandler(asyncssh.SSHServerSession):
             )
         )
         logger.info(
-            "SSH menu: user=%s connected to agent=%s hostname=%s",
+            "Gateway menu: user=%s connected to agent=%s hostname=%s",
             self._user.username, agent_id, hostname,
         )
 
@@ -623,12 +439,12 @@ class MenuSessionHandler(asyncssh.SSHServerSession):
                 if done:
                     self._chan.exit(exit_code or 0)
             except Exception:
-                logger.error("SSH menu output_cb error", exc_info=True)
+                logger.error("Gateway menu output_cb error", exc_info=True)
 
         try:
             await self._term.start(output_cb)
         except Exception as e:
-            logger.error("SSH menu: failed to start terminal: %s", e, exc_info=True)
+            logger.error("Gateway menu: failed to start terminal: %s", e, exc_info=True)
             await self._write(f"\r\n\x1b[31mFailed to connect: {e}\x1b[0m\r\n")
             self._state = "agent"
             await self._show_agents()
@@ -788,7 +604,7 @@ class MenuSessionHandler(asyncssh.SSHServerSession):
 
     async def _show_help(self):
         lines = [
-            "\r\n\x1b[1mTactical RMM SSH Gateway - Help\x1b[0m\r\n",
+            "\r\n\x1b[1mTactical RMM Gateway - Help\x1b[0m\r\n",
             "\x1b[2m----------------------------------------\x1b[0m\r\n",
             "",
             "  \x1b[1mNavigation\x1b[0m",
