+++ api/tacticalrmm/tacticalrmm/gateway/server.py
@@ -8,7 +8,8 @@ from asgiref.sync import sync_to_async
 from django.conf import settings
 
 from .audit import _audit_session_failed
-from .handlers import MenuSessionHandler, SSHSessionHandler
+from .handlers import DirectSessionHandler
+from .menu import MenuSessionHandler
 from .rate_limiter import _rate_limiter
 from .utils import _fingerprint, _lookup_key, _resolve_and_check
 
@@ -21,7 +22,7 @@ def get_active_connections() -> int:
     return _active_connections
 
 
-class SSHAgentServer(asyncssh.SSHServer):
+class GatewayServer(asyncssh.SSHServer):
     def __init__(self):
         self._auth_user = None
         self._auth_agent = None
@@ -42,7 +43,7 @@ class SSHAgentServer(asyncssh.SSHServer):
             self._remote_ip = peer[0] if peer else ""
             self._client_version = conn.get_extra_info("client_version", "")
             logger.info(
-                "SSH connection from %s client=%s",
+                "Gateway connection from %s client=%s",
                 self._remote_ip, self._client_version,
             )
         except Exception:
@@ -58,7 +59,7 @@ class SSHAgentServer(asyncssh.SSHServer):
             global _active_connections
             _active_connections -= 1
         if exc:
-            logger.error("SSH connection from %s lost: %s", self._remote_ip, exc)
+            logger.error("Gateway connection from %s lost: %s", self._remote_ip, exc)
 
     async def begin_auth(self, username):
         self._auth_user = None
@@ -66,7 +67,7 @@ class SSHAgentServer(asyncssh.SSHServer):
         self._is_menu = False
         self._session_id = uuid.uuid4().hex
         if self._remote_ip and not _rate_limiter.allow(self._remote_ip):
-            logger.warning("SSH: rate limit exceeded for %s", self._remote_ip)
+            logger.warning("Gateway: rate limit exceeded for %s", self._remote_ip)
             return False
         return True
 
@@ -89,13 +90,13 @@ class SSHAgentServer(asyncssh.SSHServer):
 
             ssh_key = await _lookup_key(fp)
             if ssh_key is None:
-                logger.warning("SSH: unknown key %s from %s", fp, self._remote_ip)
+                logger.warning("Gateway: unknown key %s from %s", fp, self._remote_ip)
                 return False
 
             user = ssh_key.user
             if not user.is_active or user.block_dashboard_login:
                 logger.warning(
-                    "SSH: inactive user %s from %s", user.username, self._remote_ip
+                    "Gateway: inactive user %s from %s", user.username, self._remote_ip
                 )
                 return False
 
@@ -107,7 +108,7 @@ class SSHAgentServer(asyncssh.SSHServer):
                 role = await sync_to_async(user.get_and_set_role_cache)()
                 if not user.is_superuser and not (role and getattr(role, "can_use_terminal", False)):
                     logger.warning(
-                        "SSH: menu denied for user=%s from %s",
+                        "Gateway: menu denied for user=%s from %s",
                         user.username, self._remote_ip,
                     )
                     return False
@@ -115,7 +116,7 @@ class SSHAgentServer(asyncssh.SSHServer):
                 self._auth_agent = None
                 self._is_menu = True
                 logger.info(
-                    "SSH: menu session user=%s key=%s from %s",
+                    "Gateway: menu session user=%s key=%s from %s",
                     user.username, ssh_key.name, self._remote_ip,
                 )
                 return True
@@ -123,7 +124,7 @@ class SSHAgentServer(asyncssh.SSHServer):
             agent = await _resolve_and_check(user, username)
             if agent is None:
                 logger.warning(
-                    "SSH: denied user=%s agent=%s from %s",
+                    "Gateway: denied user=%s agent=%s from %s",
                     user.username, username, self._remote_ip,
                 )
                 asyncio.create_task(
@@ -141,7 +142,7 @@ class SSHAgentServer(asyncssh.SSHServer):
 
             if agent.status != "online":
                 logger.warning(
-                    "SSH: denied user=%s agent=%s from %s reason=agent_%s",
+                    "Gateway: denied user=%s agent=%s from %s reason=agent_%s",
                     user.username, username, self._remote_ip, agent.status,
                 )
                 asyncio.create_task(
@@ -160,12 +161,12 @@ class SSHAgentServer(asyncssh.SSHServer):
             self._auth_user = user
             self._auth_agent = agent
             logger.info(
-                "SSH: auth success user=%s agent=%s key=%s from %s",
+                "Gateway: auth success user=%s agent=%s key=%s from %s",
                 user.username, username, ssh_key.name, self._remote_ip,
             )
             return True
         except Exception as e:
-            logger.error("SSH: validate_public_key error: %s", e, exc_info=True)
+            logger.error("Gateway: validate_public_key error: %s", e, exc_info=True)
             return False
 
     def session_requested(self):
@@ -182,7 +183,7 @@ class SSHAgentServer(asyncssh.SSHServer):
                 )
             if not self._auth_user or not self._auth_agent:
                 return None
-            handler = SSHSessionHandler(
+            handler = DirectSessionHandler(
                 self._auth_user, self._auth_agent, self._session_id, self._remote_ip,
                 client_version=self._client_version,
                 ssh_key_name=self._ssh_key_name,
@@ -191,18 +192,18 @@ class SSHAgentServer(asyncssh.SSHServer):
             )
             return handler
         except Exception as e:
-            logger.error("SSH: session_requested error: %s", e, exc_info=True)
+            logger.error("Gateway: session_requested error: %s", e, exc_info=True)
             return None
 
 
-async def start_ssh_server(host=None, port=None):
+async def start_gateway(host=None, port=None):
     if host is None:
         host = getattr(settings, "SSH_SERVER_HOST", "0.0.0.0")
     if port is None:
         port = getattr(settings, "SSH_SERVER_PORT", 2222)
 
     if not isinstance(port, int) or port < 1 or port > 65535:
-        raise ValueError(f"Invalid SSH port: {port}")
+        raise ValueError(f"Invalid gateway port: {port}")
 
     login_timeout = getattr(settings, "SSH_LOGIN_TIMEOUT", 30)
     if login_timeout < 5:
@@ -228,7 +229,7 @@ async def start_ssh_server(host=None, port=None):
     server_host_keys = [host_key_path]
 
     server = await asyncssh.create_server(
-        SSHAgentServer,
+        GatewayServer,
         host,
         port,
         server_host_keys=server_host_keys,
@@ -243,7 +244,7 @@ async def start_ssh_server(host=None, port=None):
         reuse_address=True,
     )
     logger.info(
-        "SSH gateway listening on %s:%s (timeout=%d keepalive=%d/%d)",
+        "Gateway listening on %s:%s (timeout=%d keepalive=%d/%d)",
         host, port, login_timeout, keepalive_interval, keepalive_count_max,
     )
     return server
